time_tracking/requirements.txt
Django==3.2.8
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
factory-boy==3.2.1
pytest==6.2.5
pytest-django==4.5.2
pytest-factoryboy==2.1.0
